pyproject.toml
[tool.poetry]
name = "alacritty"
version = "0.1.0"
description = ""
authors = ["raushanraja <vraushanraja@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
inquirer = "^3.1.3"
pyyaml = "^6.0.1"
toml = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
